+++ R/variable-type.R
@@ -51,6 +51,11 @@ has.categories <- function (x) {
     return(x %in% c("categorical_array", "multiple_response", "categorical"))
 }
 
+is.subvariable <- function (x) {
+    ## TODO: server should support a better way of determining this
+    grepl("subvariables", self(x))
+}
+
 CASTABLE_TYPES <- c("numeric", "text", "categorical") ## Add datetime when server supports
 
 #' Change the type of Crunch variables