+++ tests/testthat/test-append-debug.R
@@ -47,7 +47,7 @@ with_test_authentication({
             expect_error(
                 expect_message(appendDataset(part1, part2),
                     "Result URL"),
-                "Subvariable mr_1 cannot be bound to both arrays CA2 and CA1.")
+                "Subvariable mr_1 cannot be bound to both arrays 'CA2' and 'CA1'.")
         })
         part1 <- cleanseBatches(part1)