+++ NEWS.md
@@ -1,5 +1,11 @@
-### crunch 1.13.5 (under development)
+### crunch 1.13.7 (under development)
+* `flattenOrder` to strip out nested groups from an order
+
+### crunch 1.13.6
 * More utility functions for working with order objects: `dedupeOrder`, `removeEmptyGroups`
+* `appendDataset` can now append a subset of variables
+* Update to changes in the dataset version API
+* Fix bug in assigning NA to an array subvariable that didn't already have the "No Data" category
 
 ### crunch 1.13.4
 * `flipArrays` function to generate derived views of array subvariables